+++ Changelog.md
@@ -22,6 +22,7 @@
   * 'help' plugin updated.
   * all api from ybot.erl beside start/stop were removed.
   * ybot_shell added for testing Ybot from erlang shell.
+  * reconnect_timeout option added to flowdock transport.
 
 ## Ybot-0.3.2 --> released